+++ CHANGES.rst
@@ -23,6 +23,9 @@
 - Reduce the default number of attempts to 4 (one attempt and 3
   retries). See `issue 35 <https://github.com/NextThought/nti.transactions/issues/35>`_.
 
+- Make the transaction loop emit more metrics. See `issue 31
+  <https://github.com/NextThought/nti.transactions/issues/31>`_.
+
 3.0.0 (2019-09-06)
 ==================